requirements.txt
htmltools 
ipywidgets
jinja2
jupyter_bokeh
marvel 
matplotlib
pandas
plotly 
plotnine
pyodide-py
python-dotenv
requests
rsconnect-python
shiny 
shinyswatch 
shinywidgets 
